Fixing GL ES problems + cleanup parts

+++ Assets/Shaders/level.fs
@@ -1,4 +1,3 @@
-
 out vec4 FragColor;
 
 in vec2 aTexCoord;
@@ -8,7 +7,7 @@ uniform int textureID[540];
 uniform sampler2D textureA;
 
 void main() {
-	vec2 pTexCoord = aTexCoord + vec2((textureID[instanceID] / 6) / 11.0, -(textureID[instanceID] % 6) / 6.0);
+	vec2 pTexCoord = aTexCoord + vec2(float(textureID[instanceID] / 6) / 11.0, -float(textureID[instanceID] % 6) / 6.0);
 
 	FragColor = texture(textureA, pTexCoord);
 }
